Sample Data file and train a model
I Create a sample data for "Traffic Accident" between 2000 and 2010 and I use the data from 2000 and 2008 as "Train data" and I use 2009 and 2010 as "test data". Also, I choose to use "Random Forest Model" to train the data. It could be an overview about what we should do. Actually the model is not very accurate but I hope it could be useful.